[BUGFIX] Fix "All valid files" bug
Fix issue #54:
After the file dialog is used once and a file is selected with the
option "All valid files," the next file dialog opened will have the
first extension selected, but will allow all files to be selected. The
problem persists on future file dialogs until another extension is
selected on the drop down.

The fix was just adding the line "show_all_valid_files = false;" to
closeDialog()